Fix FIONBIO type

+++ src/unix/newlib/mod.rs
@@ -406,7 +406,7 @@ pub const SEEK_SET: ::c_int = 0;
 pub const SEEK_CUR: ::c_int = 1;
 pub const SEEK_END: ::c_int = 2;
 
-pub const FIONBIO: ::c_int = 1;
+pub const FIONBIO: ::c_ulong = 1;
 pub const FIOCLEX: ::c_ulong = 0x20006601;
 
 pub const S_BLKSIZE: ::mode_t = 1024;
